모델의 저장 및 복원을 위한 체크포인트 등에 관한 리팩토링 #78
